Updating no songs yet page to match the mocks
